images unpack: Add initial support for WIC images
Add support to unpack Torizon OS .wic images. The image has to have a
partition named after the value of arg 'wic_rootfs_label' (default:
otaroot) that has the sysroot contents, namely the 'boot' and 'ostree'
directories.

WIC images don't create a directory named 'tezi' in the storage
directory when unpacked.

- Add the following function in backend/common.py:

  unpacked_image_type(storage_dir): Check if the current unpacked
  image is of type 'tezi' or 'wic' by looking if storage_dir has a
  directory named tezi.

+++ tcbuilder/backend/images.py
@@ -18,6 +18,7 @@ import urllib.request
 from zipfile import ZipFile
 from tempfile import TemporaryDirectory
 
+import guestfs
 import paramiko
 
 from tcbuilder.backend.common import (get_rootfs_tarball, get_tar_compress_program_options,
@@ -248,6 +249,36 @@ def unpack_local_image(image_dir, sysroot_dir):
     os.unlink(tarfile)
 
 
+def unpack_local_wic_image(image_dir, sysroot_dir, wic_rootfs_label):
+    """Extract the root fs from the image into the sysroot directory"""
+    try:
+        gfs = guestfs.GuestFS(python_return_dict=True)
+        # Add input image
+        gfs.add_drive_opts(image_dir, format="raw", readonly=1)
+        # Launch libguestfs backend
+        gfs.launch()
+        if len(gfs.list_partitions()) < 1:
+            raise TorizonCoreBuilderError(
+                "Image doesn't have any partitions or it's not a valid WIC image. Aborting.")
+
+        # Get partition number from ext4 fs called wic_rootfs_label in disk image (.wic)
+        rootfs_partition = gfs.findfs_label(wic_rootfs_label)
+        log.info(f"'{wic_rootfs_label}' partition found: {rootfs_partition}")
+        gfs.mount_ro(rootfs_partition, "/")
+        log.info("Unpacking WIC image. This may take a few minutes...")
+        gfs.copy_out("/", sysroot_dir)
+        gfs.shutdown()
+        gfs.close()
+    except RuntimeError as gfserr:
+        if gfs:
+            gfs.close()
+        if f"unable to resolve 'LABEL={wic_rootfs_label}'" in str(gfserr):
+            raise TorizonCoreBuilderError(
+                f"Filesystem with label '{wic_rootfs_label}' not found in image. Aborting.")
+
+        raise TorizonCoreBuilderError(f"guestfs: {str(gfserr)}")
+
+
 def _make_tezi_extract_dir(tezi_dir):
     """Create target directory where to extract the tezi image"""
     extract_dir = tezi_dir + '.tmp'
@@ -257,7 +288,9 @@ def _make_tezi_extract_dir(tezi_dir):
     return extract_dir
 
 
-def import_local_image(image_dir, tezi_dir, src_sysroot_dir, src_ostree_archive_dir):
+# pylint: disable=too-many-locals
+def import_local_image(image_dir, tezi_dir, src_sysroot_dir, src_ostree_archive_dir,
+                       wic_rootfs_label="otaroot"):
     """Import local Toradex Easy Installer image
 
     Import local Toradex Easy installer image to be customized. Assuming an
@@ -297,16 +330,20 @@ def import_local_image(image_dir, tezi_dir, src_sysroot_dir, src_ostree_archive_
             file.extractall(extract_dir)
             image_dir = extract_dir
 
-    log.info("Copying Toradex Easy Installer image.")
-    log.debug(f"Copy directory {image_dir} -> {tezi_dir}.")
-    shutil.copytree(image_dir, tezi_dir)
+    if image_dir.lower().endswith(".wic"):
+        log.info("Mounting WIC image.")
+        unpack_local_wic_image(image_dir, src_sysroot_dir, wic_rootfs_label)
+    else:
+        log.info("Copying Toradex Easy Installer image.")
+        log.debug(f"Copy directory {image_dir} -> {tezi_dir}.")
+        shutil.copytree(image_dir, tezi_dir)
 
-    # Get rid of the extraction directory (if we created one).
-    if extract_dir is not None:
-        shutil.rmtree(extract_dir)
+        # Get rid of the extraction directory (if we created one).
+        if extract_dir is not None:
+            shutil.rmtree(extract_dir)
 
-    log.info("Unpacking TorizonCore Toradex Easy Installer image.")
-    unpack_local_image(tezi_dir, src_sysroot_dir)
+        log.info("Unpacking TorizonCore Toradex Easy Installer image.")
+        unpack_local_image(tezi_dir, src_sysroot_dir)
 
     src_sysroot = ostree.load_sysroot(src_sysroot_dir)
     csum, _ = ostree.get_deployment_info_from_sysroot(src_sysroot)
@@ -319,9 +356,14 @@ def import_local_image(image_dir, tezi_dir, src_sysroot_dir, src_ostree_archive_
     ostree.pull_local_refs(repo, src_ostree_dir, refs=target_refs, remote="torizon")
     metadata, _, _ = ostree.get_metadata_from_checksum(src_sysroot.repo(), csum)
 
-    log.info("Unpacked OSTree from Toradex Easy Installer image:")
+    if os.path.exists(tezi_dir):
+        log.info("Unpacked OSTree from Toradex Easy Installer image:")
+    else:
+        log.info("Unpacked OSTree from WIC image:")
+
     log.info(f"  Commit checksum: {csum}".format(csum))
     log.info(f"  TorizonCore Version: {metadata['version']}")
+# pylint: enable=too-many-locals
 
 
 def prov_check_provdata_presence(input_dir):
